Split up backend files
