+++ client/web/src/enterprise/campaigns/preview/list/ChangesetSpecList.tsx
@@ -3,7 +3,7 @@ import * as H from 'history'
 import { ThemeProps } from '../../../../../../shared/src/theme'
 import { FilteredConnection, FilteredConnectionQueryArguments } from '../../../../components/FilteredConnection'
 import { ChangesetApplyPreviewFields, Scalars } from '../../../../graphql-operations'
-import { queryChangesetSpecs as _queryChangesetSpecs, queryChangesetSpecFileDiffs } from '../backend'
+import { queryChangesetSpecs as _queryChangesetSpecs, queryChangesetSpecFileDiffs } from './backend'
 import { ChangesetSpecNode, ChangesetSpecNodeProps } from './ChangesetSpecNode'
 import { ChangesetSpecListHeader } from './ChangesetSpecListHeader'
 import { EmptyChangesetSpecListElement } from './EmptyChangesetSpecListElement'
