+++ client/web/src/enterprise/campaigns/preview/list/ChangesetSpecNode.tsx
@@ -4,7 +4,7 @@ import { ChangesetApplyPreviewFields } from '../../../../graphql-operations'
 import { HiddenChangesetSpecNode } from './HiddenChangesetSpecNode'
 import { VisibleChangesetSpecNode } from './VisibleChangesetSpecNode'
 import { ThemeProps } from '../../../../../../shared/src/theme'
-import { queryChangesetSpecFileDiffs } from '../backend'
+import { queryChangesetSpecFileDiffs } from './backend'
 
 export interface ChangesetSpecNodeProps extends ThemeProps {
     node: ChangesetApplyPreviewFields
