+++ client/web/src/enterprise/campaigns/preview/list/VisibleChangesetSpecNode.tsx
@@ -2,7 +2,7 @@ import * as H from 'history'
 import React, { useCallback, useState } from 'react'
 import { VisibleChangesetApplyPreviewFields, VisibleChangesetSpecFields } from '../../../../graphql-operations'
 import { ThemeProps } from '../../../../../../shared/src/theme'
-import { queryChangesetSpecFileDiffs as _queryChangesetSpecFileDiffs } from '../backend'
+import { queryChangesetSpecFileDiffs as _queryChangesetSpecFileDiffs } from './backend'
 import { Link } from '../../../../../../shared/src/components/Link'
 import { DiffStat } from '../../../../components/diff/DiffStat'
 import { ChangesetSpecAction } from './ChangesetSpecAction'
